workshops/o11y-vector-ai/data-generator/requirements.txt
httpx==0.26.0
faker==22.5.1
pyyaml==6.0.1
python-dotenv==1.0.0
tenacity==8.2.3
